# Faster algorithm for ordered sampling with replacement