--- a/src/sampling.jl
+++ b/src/sampling.jl
@@ -522,11 +522,7 @@ function sample!(rng::AbstractRNG, a::AbstractArray, x::AbstractArray;
 
     if replace  # with replacement
         if ordered
-            if k > 10
-                uniform_orderstat_sample!(rng, a, x)
-            else
-                sample_ordered!(direct_sample!, rng, a, x)
-            end
+            uniform_orderstat_sample!(rng, a, x)
         else
             direct_sample!(rng, a, x)
         end