--- a/src/sampling.jl
+++ b/src/sampling.jl
@@ -525,7 +525,7 @@ function sample!(rng::AbstractRNG, a::AbstractArray, x::AbstractArray;
             if k <= 10
                 sample_ordered!(direct_sample!, rng, a, x)
             else
-                uniform_orderdist_sample!(rng, a, x)
+                uniform_orderstat_sample!(rng, a, x)
             end        
         else
             direct_sample!(rng, a, x)