This is based on a classical result for example described here https://stats.stackexchange.com/questions/348358/a-fast-uniform-order-statistic-generator (and in the reference of the most upvoted answer). I wasn't able to find a reference describing its modification for sampling a finite population, but I adapted it to such a case.

In particular, the performance increase is substantial, more than 5 times faster when the performance increase stabilize, e.g.

This PR:

```julia
julia> using StatsBase, BenchmarkTools

julia> a = [1:1000;];

julia> @btime sample($a, 10^2, ordered=true);
  453.017 ns (2 allocations: 1.75 KiB)

julia> @btime sample($a, 10^6, ordered=true);
  3.491 ms (4 allocations: 15.26 MiB)
```

Main:

```julia
julia> using StatsBase, BenchmarkTools

julia> a = [1:1000;];

julia> @btime sample($a, 10^2, ordered=true);
  1.564 μs (2 allocations: 1.75 KiB)

julia> @btime sample($a, 10^6, ordered=true);
  21.273 ms (4 allocations: 15.26 MiB)
```

the switching point between this algorithm and the one implemented in main is set at `k=10` because I found that empirically at that point the timings were almost equal.

Numerically it should be stable enough, but let me know what you think