@Tortar 2024-01-01
Test that is failing is using the previous algorithm at lines 83-84 of sampling.jl

```julia
83    aa = Int.(sample(r, 10; ordered=true))
84    check_sample_wrep(aa, (3, 12), 0; ordered=true, rev=rev)
```

so I think it is unrelated right?