@Tortar 2024-04-12
gentle bump, since #927 was merged, what about taking a look at another speed-up? :D